functions/requirements.txt
functions-framework==3.*
firebase-admin==6.5.0
google-cloud-aiplatform==1.70.0
requests==2.32.3
python-dotenv==1.0.1